media-dashboard/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
nltk==3.9.1
requests==2.32.3
python-dotenv==1.0.1
